complete lab2

+++ lab2/DebugPractice/DebugExercise1.java
@@ -3,9 +3,9 @@
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+    public static int divideThenRound(double top, double bottom) {
+        double quotient = top / bottom;
+        int result = (int) Math.round(quotient);
         return result;
     }
